requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
Django==5.0.2
django-cors-headers==4.3.1
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
mysqlclient==2.2.4
packaging==23.2
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.2
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
typing_extensions==4.10.0
uritemplate==4.1.1
